requirements.txt
absl-py==1.4.0
aiohttp==3.8.3
aiosignal==1.3.1
antlr4-python3-runtime==4.8
appdirs==1.4.4
astunparse==1.6.3
async-timeout==4.0.2
attrs==19.3.0
audioread==3.0.0
Automat==0.8.0
bitarray==2.6.2
blinker==1.4
blis==0.7.9
cachetools==5.3.0
catalogue==2.0.8
certifi==2019.11.28
cffi==1.15.1
chardet==3.0.4
charset-normalizer==2.1.1
click==8.1.3
cloud-init==22.4.2
colorama==0.4.3
command-not-found==0.3
confection==0.0.4
configobj==5.0.6
constantly==15.1.0
contourpy==1.0.7
cryptography==2.8
cycler==0.11.0
cymem==2.0.7
Cython==0.29.33
datasets==2.9.0
dbus-python==1.2.16
decorator==4.4.2
dill==0.3.6
distro==1.4.0
distro-info===0.23ubuntu1
editdistance==0.6.2
einops==0.6.0
en-core-web-sm==3.5.0
entrypoints==0.3
filelock==3.9.0
Flask==2.2.2
flatbuffers==23.1.21
fonttools==4.38.0
frozenlist==1.3.3
fsspec==2023.1.0
ftfy==6.0.3
gast==0.4.0
google-auth==2.16.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.51.1
h5py==3.8.0
httplib2==0.14.0
huggingface-hub==0.11.1
hydra-core==1.0.7
hyperlink==19.0.0
idna==2.8
imageio==2.25.0
imageio-ffmpeg==0.4.8
importlib-metadata==6.0.0
importlib-resources==5.10.2
incremental==16.10.1
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.2.0
jsonpatch==1.22
jsonpointer==2.0
jsonschema==3.2.0
keras==2.11.0
keyring==18.0.1
kiwisolver==1.4.4
langcodes==3.3.0
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
Levenshtein==0.20.9
libclang==15.0.6.1
librosa==0.9.2
lightning-utilities==0.6.0.post0
llvmlite==0.39.1
lxml==4.9.2
Markdown==3.4.1
MarkupSafe==2.1.2
matplotlib==3.6.3
more-itertools==4.2.0
moviepy==1.0.3
multidict==6.0.4
multiprocess==0.70.14
murmurhash==1.0.9
netifaces==0.10.4
networkx==3.0
nltk==3.8.1
numba==0.56.4
numpy==1.23.2
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
oauthlib==3.1.0
omegaconf==2.0.6
opencv-python==4.7.0.68
opt-einsum==3.3.0
packaging==23.0
pandas==1.5.3
pathy==0.10.1
pexpect==4.6.0
Pillow==9.4.0
pooch==1.6.0
portalocker==2.7.0
preshed==3.0.8
proglog==0.1.10
protobuf==3.19.6
pyarrow==10.0.1
pyasn1==0.4.2
pyasn1-modules==0.2.1
pycocoevalcap==1.2
pycocotools==2.0.4
pycparser==2.21
pydantic==1.10.4
PyGObject==3.36.0
PyHamcrest==1.9.0
PyJWT==1.7.1
pymacaroons==0.13.0
PyNaCl==1.3.0
pyOpenSSL==19.0.0
pyparsing==3.0.9
pyparted==3.11.2
pypinyin==0.47.1
pyrsistent==0.15.5
pyserial==3.4
pytesseract==0.3.10
python-apt==2.0.0+ubuntu0.20.4.8
python-dateutil==2.8.2
python-debian===0.1.36ubuntu1
python-dotenv==0.21.1
python-Levenshtein==0.20.9
pytorch-lightning==1.9.0
pytz==2022.7.1
PyWavelets==1.4.1
PyYAML==5.3.1
rapidfuzz==2.13.7
regex==2022.10.31
requests==2.22.0
requests-oauthlib==1.3.1
requests-unixsocket==0.2.0
resampy==0.4.2
responses==0.18.0
rouge-score==0.1.2
rsa==4.9
sacrebleu==2.3.1
sacremoses==0.0.53
scikit-image==0.19.3
scikit-learn==1.2.1
scipy==1.10.0
SecretStorage==2.3.1
sentencepiece==0.1.97
service-identity==18.1.0
session-info==1.0.0
simplejson==3.16.0
six==1.14.0
smart-open==6.3.0
sos==4.4
soundfile==0.11.0
spacy==3.5.0
spacy-legacy==3.0.12
spacy-loggers==1.0.4
srsly==2.4.5
ssh-import-id==5.10
stdlib-list==0.8.0
systemd-python==234
tabulate==0.9.0
termcolor==2.2.0
tesseract==0.1.3
thinc==8.1.7
threadpoolctl==3.1.0
tifffile==2023.2.2
timm==0.6.12
tokenizers==0.13.2
torch==1.13.1
torchmetrics==0.11.0
torchvision==0.14.1
tqdm==4.64.1
transformers==4.18.0.dev0
Twisted==18.9.0
typer==0.7.0
typing-extensions==4.4.0
ubuntu-advantage-tools==27.12
ufw==0.36
unattended-upgrades==0.1
urllib3==1.26.14
wadllib==1.3.3
WALinuxAgent==2.2.46
wasabi==1.1.1
wcwidth==0.2.6
Werkzeug==2.2.2
wrapt==1.14.1
xxhash==3.2.0
yarl==1.8.2
zhconv==1.4.3
zipp==3.11.0
zope.interface==4.7.1
